refactor OTFCompileExecutors -> ModularExecutor
