+++ docs/development/ADRs/0015-Test_Exclusion_Matrices.md
@@ -49,9 +49,9 @@ The following backend processors are defined:
 ```python
 DACE_CPU = "dace_iterator.run_dace_cpu"
 DACE_GPU = "dace_iterator.run_dace_gpu"
-GTFN_CPU = "otf_compile_executor.run_gtfn"
-GTFN_CPU_IMPERATIVE = "otf_compile_executor.run_gtfn_imperative"
-GTFN_CPU_WITH_TEMPORARIES = "otf_compile_executor.run_gtfn_with_temporaries"
+GTFN_CPU = "gt4py.next.program_processors.runners.gtfn.run_gtfn"
+GTFN_CPU_IMPERATIVE = "gt4py.next.program_processors.runners.gtfn.run_gtfn_imperative"
+GTFN_CPU_WITH_TEMPORARIES = "gt4py.next.program_processors.runners.gtfn.run_gtfn_with_temporaries"
 GTFN_GPU = "gt4py.next.program_processors.runners.gtfn.run_gtfn_gpu"
 ```
