+++ src/gt4py/next/program_processors/runners/double_roundtrip.py
@@ -16,9 +16,9 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any
 
-import gt4py.next.program_processors.otf_compile_executor as otf_compile_executor
 import gt4py.next.program_processors.processor_interface as ppi
-import gt4py.next.program_processors.runners.roundtrip as roundtrip
+from gt4py.next import backend as next_backend
+from gt4py.next.program_processors.runners import roundtrip
 
 
 if TYPE_CHECKING:
@@ -30,7 +30,7 @@ def executor(program: itir.FencilDefinition, *args: Any, **kwargs: Any) -> None:
     roundtrip.execute_roundtrip(program, *args, dispatch_backend=roundtrip.executor, **kwargs)
 
 
-backend = otf_compile_executor.OTFBackend(
+backend = next_backend.Backend(
     executor=executor,
     allocator=roundtrip.backend.allocator,
 )
