+++ tests/next_tests/integration_tests/feature_tests/ffront_tests/test_temporaries_with_sizes.py
@@ -16,9 +16,9 @@ import pytest
 from numpy import int32, int64
 
 from gt4py import next as gtx
-from gt4py.next import common
+from gt4py.next import backend, common
 from gt4py.next.iterator.transforms import LiftMode, apply_common_transforms
-from gt4py.next.program_processors import otf_compile_executor
+from gt4py.next.program_processors import modular_executor
 from gt4py.next.program_processors.runners.gtfn import run_gtfn_with_temporaries
 
 from next_tests.integration_tests import cases
@@ -38,8 +38,8 @@ from next_tests.toy_connectivity import Cell, Edge
 
 @pytest.fixture
 def run_gtfn_with_temporaries_and_symbolic_sizes():
-    return otf_compile_executor.OTFBackend(
-        executor=otf_compile_executor.OTFCompileExecutor(
+    return backend.Backend(
+        executor=modular_executor.ModularExecutor(
             name="run_gtfn_with_temporaries_and_sizes",
             otf_workflow=run_gtfn_with_temporaries.executor.otf_workflow.replace(
                 translation=run_gtfn_with_temporaries.executor.otf_workflow.translation.replace(
